Outline
Clean Outline to begin project.

(I am in no way confident in this. The group projects from the week meant I never got to learn this on my own which is the only way I'm able to retain these things. I am flying about 80% blind in objective-C and I expect this will take the entire weekend.