Added Models
Added my artist models was having a hard issue when working with the NSJSONSerialization. I was not sure what was the correct method in accessing the Artist Model. I imported "SAEArtist.h" to my  NSJSONSerialization+SAEArtist.h but when creating the method in NSJSONSerialization+SAEArtist.m i found the only way to access the Artist Model was to change the implementation to SAEArtist (NSJSONSerialization). This allowed me to save to the Artist Model.